move Message XML tests to own file
Also, fix a DD file path to use system constant.

 #585
